// src/app/shared/model/InstitucionEducativa.ts
/**
 * <Derechos_Reservados>
 * 
 * Aplicacion		:SISPAE 
 * 
 * Autor			:TiGlobal SAS y SoftManagement
 * 
 * Generacion		:Este archivo es generado automaticamente mediante generador GeneraApp.
 * 
 * Ano			    :2022
 * 
 * Arquitectura	    :Modelo Base de Angular para Arquitectura Microservicios de la Entidad InstitucionEducativa
 * 
 * Capa			    :SISPAE-Front 
 * 
 * </Derechos_Reservados>
 */


/**
 * Definicion de Interface  con los atributos del objeto, desprovistas de inicialización y funcionalidad,
 *
 *  La interfaz es el contrato entre el mundo exterior y la clase
 */
export interface InstitucionEducativaInterface {

  id: number;
  iD_Rector: number;
  siD_Rector: string;
  iD_ETC: number;
  siD_ETC: string;
  iD_ET: number;
  siD_ET: string;
  iD_DiviPola: number;
  siD_DiviPola: string;
  codigoDane: string;
  nombre: string;
  correo: string;
  auditoria: string;

  // atributos para gestión de auditoria del objeto
  _ippublica: string;
  _nombremaquina: string;
  _usuario: string;
  _ipdetrasproxy: string;
  _browser: string;
  _accion: string;
  _sessionid: string;
  _XMLAuditoria: string;
  // atributos adicionales genericos para gestión del objeto
  isValid: boolean;
  isSelected: boolean;
  completed: boolean;

}

/**
 * Clase Modelo constructor que implementa interface, esta clase tiene que declarar todos los atributos y métodos que dice la interfaz
 *
 * En caso de usar clases, tus nuevos objetos deben ser creados con la palabra "new"
 */
export class InstitucionEducativaModel implements InstitucionEducativaInterface {
  constructor(

    public id: number,
    public iD_Rector: number,
    public siD_Rector: string,
    public iD_ETC: number,
    public siD_ETC: string,
    public iD_ET: number,
    public siD_ET: string,
    public iD_DiviPola: number,
    public siD_DiviPola: string,
    public codigoDane: string,
    public nombre: string,
    public correo: string,
    public auditoria: string,

    // atributos para gestión de auditoria del objeto
    public _ippublica: string,
    public _nombremaquina: string,
    public _usuario: string,
    public _ipdetrasproxy: string,
    public _browser: string,
    public _accion: string,
    public _sessionid: string,
    public _XMLAuditoria: string,
    // atributos adicionales genericos para gestión del objeto
    public isValid: boolean = true,
    public isSelected: boolean = false,
    public completed: boolean = false

  ) { }
}